pyproject.toml
[project]
name = "dlsidecar"
version = "0.1.0"
description = "Data Lake as a Service — sidecar container for enterprise EKS deployments"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "duckdb>=1.1.0",
    "pyarrow>=17.0.0",
    "sqlglot>=25.0.0",
    "trino>=0.329.0",
    "pyiceberg>=0.7.0",
    "pyhive[hive_pure_sasl]>=0.7.0",
    "thrift>=0.20.0",
    "prometheus-client>=0.21.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.9.0",
    "structlog>=24.4.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.0",
    "streamlit>=1.38.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
kafka = [
    "confluent-kafka>=2.5.0",
    "fastavro>=1.9.0",
]
snowflake = ["snowflake-connector-python>=3.12.0"]
databricks = ["databricks-sql-connector>=3.4.0"]
postgres = ["psycopg[binary]>=3.2.0"]
mysql = ["mysql-connector-python>=9.0.0"]
mongo = ["pymongo>=4.9.0"]
azure = ["azure-storage-blob>=12.22.0"]
gcs = ["gcsfs>=2024.9.0"]
delta = ["deltalake>=0.20.0"]
motherduck = ["duckdb>=1.1.0"]
flight = ["pyarrow>=17.0.0"]
all = [
    "dlsidecar[kafka,snowflake,databricks,postgres,mysql,mongo,azure,gcs,delta,motherduck,flight]",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "moto[s3]>=5.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dlsidecar"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
